comicc_api_project/requirements.txt
Flask
flask-cors
Pillow
rarfile
waitress
